sql. Introduce DUMMY_OPT_DEF
The set of valid options is defined in a form of opt_def structures
list. Allow for options that are validated but not copied to the target
object. This is a prerequisite for introducing sql option in indices.

Revert "sql. Introduce DUMMY_OPT_DEF"

Store truncated sql statement instead.

DUMMY_OPT_DEF resulted in a buffer overflow because opt_set() wasn't
updated.