vinyl: skip WAL records that were dumped to disk
On local recovery it can occur that a statement replayed from the WAL
was written to a run file by the Vinyl engine before restart to free up
memory. Replaying such statements is useless, besides it can result in
exceeding the memory quota during recovery, which isn't allowed as the
scheduler isn't started until recovery is complete. Before commit
a0d27684a ("vinyl: apply TX in vy_prepare and make vinyl similar to
memtx") we filtered out such statements by checking the statement LSN
before inserting it into the memory tree (in vy_commit()). The above
mentioned commit broke this check, because it moved LSN assignment after
tree insertion, but even before the commit the check was inefficient
since in the presence of secondary keys we still read the primary index
on recovery even if we didn't intend to apply the statement.

That said we need to check if the statement was applied before
committing it, in vy_replace(), vy_delete(), vy_update(), and
vy_upsert(). As we don't know the statement LSN there yet, we pass a
pointer to recovery->vclock to Engine::beginInitialRecovery(). The
vclock is incremented before applying each WAL row so we can use it to
get the LSN of a statement coming from WAL.

Closes #1900