vinyl: sort run slices in vylog
Since commit 3a6c2ff320e4 ("vinyl: sort slices by lsn on recovery")
vylog doesn't guarantee that on recovery run slices are returned in the
chronological order, instead slices are sorted by LSN recovered from
index files after reading vylog. This is conceptually wrong, because
index files are supposed to be supplementary, ideally, Vinyl should be
able to function without them. So this patch makes vylog iterate over
slices in the chronological order, as it used to. To do that, we need to
store LSN of each run in vylog. Let's reuse the 'dump_lsn' key for this.